+++ pynsee/macrodata/_load_dataset_data.py
@@ -1,9 +1,9 @@
 import pandas as pd
 from tqdm import trange
 
-from pynsee.macrodata.get_dataset_list import get_dataset_list
-from pynsee.macrodata._get_dataset_metadata import _get_dataset_metadata
-from pynsee.utils.save_df import save_df
+from ..utils.save_df import save_df
+from .get_dataset_list import get_dataset_list
+from ._get_metadata import _get_dataset_metadata
 
 
 @save_df(day_lapse_max=90)
