+++ pynsee/macrodata/get_column_title.py
@@ -4,9 +4,8 @@
 import pandas as pd
 from tqdm import trange
 
-from pynsee.macrodata.get_dataset_list import get_dataset_list
-from pynsee.macrodata._get_dataset_dimension import _get_dataset_dimension
-from pynsee.macrodata._get_dimension_values import _get_dimension_values
+from .get_dataset_list import get_dataset_list
+from ._get_dimensions import _get_dataset_dimension, _get_dimension_values
 
 
 def get_column_title(dataset=None, update=True):
