+++ tests/macrodata/test_pynsee_macrodata.py
@@ -15,10 +15,12 @@ from pynsee.macrodata._get_idbank_internal_data_harmonized import (
 from pynsee.macrodata._get_idbank_internal_data import (
     _get_idbank_internal_data,
 )
-from pynsee.macrodata._get_dataset_metadata import _get_dataset_metadata
-from pynsee.macrodata._get_dataset_dimension import _get_dataset_dimension
-from pynsee.macrodata._get_dimension_values import _get_dimension_values
-from pynsee.macrodata._download_idbank_list import _download_idbank_list
+from pynsee.macrodata._get_metadata import _get_dataset_metadata
+from pynsee.macrodata._get_dimensions import (
+    _get_dataset_dimension,
+    _get_dimension_values,
+)
+from pynsee.macrodata._download_idbanks import _download_idbank_list
 from pynsee.macrodata._get_dataset_list_internal import (
     _get_dataset_list_internal,
 )
@@ -35,6 +37,7 @@ from pynsee.macrodata.get_series_title import get_series_title
 from pynsee.macrodata.search_macrodata import search_macrodata
 
 from pynsee.utils._clean_insee_folder import _clean_insee_folder
+
 
 test_SDMX = True
